EmojiResources/requirements.txt
Pillow==8.4.0
PyYAML==5.4
unityparser==2.2.1
